requirements/base.txt
pytz==2021.1  # https://github.com/stub42/pytz
python-slugify==5.0.2  # https://github.com/un33k/python-slugify
Pillow==8.2.0  # https://github.com/python-pillow/Pillow
rcssmin==1.0.6  # https://github.com/ndparker/rcssmin
argon2-cffi==20.1.0  # https://github.com/hynek/argon2_cffi
redis==3.5.3  # https://github.com/andymccurdy/redis-py
hiredis==1.1.0  # https://github.com/redis/hiredis-py
celery==4.4.6  # pyup: < 5.0,!=4.4.7  # https://github.com/celery/celery
django-celery-beat==2.2.0  # https://github.com/celery/django-celery-beat
flower==0.9.7  # https://github.com/mher/flower

# Django
# ------------------------------------------------------------------------------
django==3.1.13  # pyup: < 3.2  # https://www.djangoproject.com/
django-environ==0.4.5  # https://github.com/joke2k/django-environ
django-model-utils==4.1.1  # https://github.com/jazzband/django-model-utils
django-allauth==0.44.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==1.12.0  # https://github.com/django-crispy-forms/django-crispy-forms
django-compressor==2.4.1  # https://github.com/django-compressor/django-compressor
django-redis==5.0.0  # https://github.com/jazzband/django-redis


# Admin
# ------------------------------------------------------------------------------
django-jet2==1.0.12
FeedParser==5.2.1
django-ipware==3.0.1
django-autocomplete-light==3.8.2
django-tinymce==3.3.0
django-filebrowser-no-grappelli==3.8.0
google-api-python-client==1.4.1
django-admin-honeypot==1.1.0


# Fileifeld upload validator
python-magic==0.4.24


# Auto import shell tool
# ------------------------------------------------------------------------------
django-shell-plus==1.1.7

# Minify HTML in django
# ------------------------------------------------------------------------------
# Todo: find the uses of these to minify wesbites and others to improve security
django-htmlmin==0.11.0
python-memcached==1.59

# List of countries
# ------------------------------------------------------------------------------
django-countries-plus==1.3.1


# Add image resize for models
# ------------------------------------------------------------------------------
django-resized==0.3.11


# Add geolocation to django app
# ------------------------------------------------------------------------------
geoip2==4.1.0
maxminddb==2.0.3


# Add medium type of read time to a post
# ------------------------------------------------------------------------------
readtime==1.1.1
